Version 1 of the project completed;
Spring Boot application with an endpoint at /forecast/{location} that returns the weather condition for the next day at the specified location. You can test the endpoint by sending a GET request to http://localhost:8080/forecast/{location}, where {location} is the location you want to get the forecast for. For example, a request to http://localhost:8080/forecast/San%20Vicente would return the weather condition for the next day in San Vicente

->Added application.yml file to Implement GitHub Action pipelines to build the .jar of the project.
->Changed the WeatherService file to return a json object